General overhaul
Included
- Removed references to pilot Role instead replaced with Captain
-Basis for two new enemies created Ramming ship and Fortress
(Ramming ship will collide with player dealing damage Fortress will is an high health enemy with increased fire rate)
-Image for Fort enemy
-Adjusted position of settings button
-New K Definitions
-Removed Textured ID's for unused ships

+++ MDP_CA1/Container.cpp
@@ -103,7 +103,7 @@ void gui::Container::SelectPrevious()
     int prev = m_selected_child;
     do
     {
-        prev = (prev + m_children.size() - 1) % m_children.size();
+        prev = static_cast<int>((prev + m_children.size() - 1) % m_children.size());
     } while (!m_children[prev]->IsSelectable());
     Select(prev);
 }
